Moves get_collider_components to physics_system
